requirements.txt
anyio==3.6.2
CacheControl==0.12.11
cachetools==5.2.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
cryptography==38.0.3
Deprecated==1.2.13
ecdsa==0.18.0
fastapi==0.85.2
firebase-admin==6.0.1
gcloud==0.18.3
google-api-core==2.10.2
google-api-python-client==2.65.0
google-auth==2.14.0
google-auth-httplib2==0.1.0
google-cloud-core==2.3.2
google-cloud-firestore==2.7.2
google-cloud-storage==2.5.0
google-crc32c==1.5.0
google-resumable-media==2.4.0
googleapis-common-protos==1.56.4
grpcio==1.50.0
grpcio-status==1.50.0
h11==0.14.0
httplib2==0.21.0
idna==3.4
jwcrypto==1.4.2
msgpack==1.0.4
oauth2client==4.1.3
proto-plus==1.22.1
protobuf==4.21.9
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pycryptodome==3.15.0
pydantic==1.10.2
PyJWT==2.6.0
pyparsing==3.0.9
Pyrebase4==4.5.0
python-dotenv==0.21.0
python-jose==3.3.0
python-jwt==4.0.0
python-multipart==0.0.5
requests==2.28.1
requests-toolbelt==0.10.1
rsa==4.9
six==1.16.0
sniffio==1.3.0
sse-starlette==1.3.3
starlette==0.20.4
typing_extensions==4.4.0
uritemplate==4.1.1
urllib3==1.26.12
uvicorn==0.19.0
wrapt==1.14.1
